requirements.txt
parsedatetime
matplotlib
discord.py
colorlog
tabulate
jishaku
asyncpg
black
